+++ contracts/p0/mixins/TradingLib.sol
@@ -274,7 +274,7 @@ library TradingLibP0 {
         view
         returns (BasketRange memory range)
     {
-        (uint192 buPriceLow, uint192 buPriceHigh) = ctx.bh.lotPrice(); // {UoA/BU}
+        (uint192 buPriceLow, uint192 buPriceHigh) = ctx.bh.price(); // {UoA/BU}
 
         // Cap ctx.basketsHeld.top
         if (ctx.basketsHeld.top > ctx.rToken.basketsNeeded()) {
@@ -303,21 +303,14 @@ library TradingLibP0 {
                 bal = bal.plus(asset.bal(address(ctx.stRSR)));
             }
 
-            {
-                // Skip over dust-balance assets not in the basket
-                (uint192 lotLow, ) = asset.lotPrice(); // {UoA/tok}
-
-                // Intentionally include value of IFFY/DISABLED collateral
-                if (
-                    ctx.bh.quantity(erc20s[i]) == 0 &&
-                    !isEnoughToSell(asset, bal, lotLow, ctx.minTradeVolume)
-                ) continue;
-            }
-
             (uint192 low, uint192 high) = asset.price(); // {UoA/tok}
-            // price() is better than lotPrice() here: it's important to not underestimate how
-            // much value could be in a token that is unpriced by using a decaying high lotPrice.
-            // price() will return [0, FIX_MAX] in this case, which is preferable.
+
+            // Skip over dust-balance assets not in the basket
+            // Intentionally include value of IFFY/DISABLED collateral
+            if (
+                ctx.bh.quantity(erc20s[i]) == 0 &&
+                !isEnoughToSell(asset, bal, low, ctx.minTradeVolume)
+            ) continue;
 
             // throughout these sections +/- is same as Fix.plus/Fix.minus and </> is Fix.gt/.lt
 
@@ -354,7 +347,7 @@ library TradingLibP0 {
 
                 // (2) Lose minTradeVolume to dust (why: auctions can return tokens)
                 // Q: Why is this precisely where we should take out minTradeVolume?
-                // A: Our use of isEnoughToSell always uses the low price (lotLow, technically),
+                // A: Our use of isEnoughToSell always uses the low price (low, technically),
                 //   so min trade volumes are always assesed based on low prices. At this point
                 //   in the calculation we have already calculated the UoA amount corresponding to
                 //   the excess token balance based on its low price, so we are already set up
@@ -453,19 +446,19 @@ library TradingLibP0 {
             // {tok} = {BU} * {tok/BU}
             uint192 needed = range.top.mul(ctx.bh.quantity(erc20s[i]), CEIL); // {tok}
             if (bal.gt(needed)) {
-                (uint192 lotLow, uint192 lotHigh) = asset.lotPrice(); // {UoA/sellTok}
-                if (lotHigh == 0) continue; // Skip worthless assets
+                (uint192 low, uint192 high) = asset.price(); // {UoA/sellTok}
+                if (high == 0) continue; // Skip worthless assets
 
                 // by calculating this early we can duck the stack limit but be less gas-efficient
                 bool enoughToSell = isEnoughToSell(
                     asset,
                     bal.minus(needed),
-                    lotLow,
+                    low,
                     ctx.minTradeVolume
                 );
 
                 // {UoA} = {sellTok} * {UoA/sellTok}
-                uint192 delta = bal.minus(needed).mul(lotLow, FLOOR);
+                uint192 delta = bal.minus(needed).mul(low, FLOOR);
 
                 // status = asset.status() if asset.isCollateral() else SOUND
                 CollateralStatus status; // starts SOUND
@@ -476,8 +469,8 @@ library TradingLibP0 {
                 if (isBetterSurplus(maxes, status, delta) && enoughToSell) {
                     trade.sell = asset;
                     trade.sellAmount = bal.minus(needed);
-                    trade.prices.sellLow = lotLow;
-                    trade.prices.sellHigh = lotHigh;
+                    trade.prices.sellLow = low;
+                    trade.prices.sellHigh = high;
 
                     maxes.surplusStatus = status;
                     maxes.surplus = delta;
@@ -487,17 +480,17 @@ library TradingLibP0 {
                 needed = range.bottom.mul(ctx.bh.quantity(erc20s[i]), CEIL); // {buyTok};
                 if (bal.lt(needed)) {
                     uint192 amtShort = needed.minus(bal); // {buyTok}
-                    (uint192 lotLow, uint192 lotHigh) = asset.lotPrice(); // {UoA/buyTok}
+                    (uint192 low, uint192 high) = asset.price(); // {UoA/buyTok}
 
                     // {UoA} = {buyTok} * {UoA/buyTok}
-                    uint192 delta = amtShort.mul(lotHigh, CEIL);
+                    uint192 delta = amtShort.mul(high, CEIL);
 
                     // The best asset to buy is whichever asset has the largest deficit
                     if (delta.gt(maxes.deficit)) {
                         trade.buy = ICollateral(address(asset));
                         trade.buyAmount = amtShort;
-                        trade.prices.buyLow = lotLow;
-                        trade.prices.buyHigh = lotHigh;
+                        trade.prices.buyLow = low;
+                        trade.prices.buyHigh = high;
 
                         maxes.deficit = delta;
                     }
@@ -512,13 +505,13 @@ library TradingLibP0 {
             uint192 rsrAvailable = rsrAsset.bal(address(ctx.bm)).plus(
                 rsrAsset.bal(address(ctx.stRSR))
             );
-            (uint192 lotLow, uint192 lotHigh) = rsrAsset.lotPrice(); // {UoA/RSR}
+            (uint192 low, uint192 high) = rsrAsset.price(); // {UoA/RSR}
 
-            if (lotHigh > 0 && isEnoughToSell(rsrAsset, rsrAvailable, lotLow, ctx.minTradeVolume)) {
+            if (high > 0 && isEnoughToSell(rsrAsset, rsrAvailable, low, ctx.minTradeVolume)) {
                 trade.sell = rsrAsset;
                 trade.sellAmount = rsrAvailable;
-                trade.prices.sellLow = lotLow;
-                trade.prices.sellHigh = lotHigh;
+                trade.prices.sellLow = low;
+                trade.prices.sellHigh = high;
             }
         }
     }