+++ test/Recollateralization.test.ts
@@ -1015,9 +1015,6 @@ describe(`Recollateralization - P${IMPLEMENTATION}`, () => {
       })
 
       it('Should not recollateralize when switching basket if all assets are UNPRICED', async () => {
-        // Set price to use lot price
-        await setOraclePrice(collateral0.address, MAX_UINT256.div(2))
-
         // Setup prime basket
         await basketHandler.connect(owner).setPrimeBasket([token1.address], [fp('1')])
 
@@ -1029,7 +1026,7 @@ describe(`Recollateralization - P${IMPLEMENTATION}`, () => {
         // Advance time post warmup period - temporary IFFY->SOUND
         await advanceTime(Number(config.warmupPeriod) + 1)
 
-        // Set to sell price = 0
+        // Set all assets to UNPRICED
         await advanceTime(Number(ORACLE_TIMEOUT.add(PRICE_TIMEOUT)))
 
         // Check state remains SOUND
@@ -1188,8 +1185,8 @@ describe(`Recollateralization - P${IMPLEMENTATION}`, () => {
           await expectRTokenPrice(rTokenAsset.address, fp('1'), ORACLE_ERROR)
         })
 
-        it('Should recollateralize correctly when switching basket - Using lot price', async () => {
-          // Set price to unpriced (will use lotPrice to size trade)
+        it('Should recollateralize correctly when switching basket', async () => {
+          // Set oracle value out-of-range
           await setOraclePrice(collateral0.address, MAX_UINT256.div(2))
 
           // Setup prime basket
@@ -1218,7 +1215,7 @@ describe(`Recollateralization - P${IMPLEMENTATION}`, () => {
             await toMinBuyAmt(sellAmt, fp('1'), fp('1')),
             6
           ).add(1)
-          // since within oracleTimeout lotPrice() should still be at 100% of original price
+          // since within oracleTimeout, price() should still be at 100% of original price
 
           await expect(facadeTest.runAuctionsForAllTraders(rToken.address))
             .to.emit(backingManager, 'TradeStarted')